FlaskApp/requirements.txt
psycopg2-binary
flask
flask-sqlalchemy
pandas
geoalchemy
flask-migrate
shapely
